Sort commands into individual command files.
